requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dataclasses==0.6
elasticsearch==7.16.0
fake-useragent==0.1.11
Flask==2.0.2
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
heroku==0.1.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-service==2.4.0
multidict==5.2.0
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
Pillow==8.4.0
pycares==4.1.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==1.5
python-socks==2.0.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
soupsieve==2.3.1typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
yarl==1.7.2
